app/requirements.txt
streamlit>=1.28.0
databricks-sql-connector>=3.0.0
databricks-sdk>=0.20.0
pandas>=1.5.0
requests>=2.28.0
